pyproject.toml
[project]
name = "apple-notes-wiki-daemon"
version = "0.1.0"
description = "Background daemon that maintains a personal wiki inside Apple Notes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "markdown>=3.6",
  "markdownify>=0.12.0",
  "openai>=1.40.0",
  "pymupdf>=1.24.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
  "watchdog>=4.0.0",
]

[tool.uv]
package = false
